+++ tests/User/UserServiceTest.php
@@ -52,6 +52,7 @@ final class UserServiceTest extends ApiTestCase
         $this->assertSame(null, $user->getBirthday());
         $this->assertNull($user->getCompanyId());
         $this->assertFalse($user->isVendor());
+        $this->assertSame('C', $user->getType());
 
         // shipping address
         $this->assertNull($user->getShippingAddress()->getTitle());
@@ -130,6 +131,7 @@ final class UserServiceTest extends ApiTestCase
         $this->assertSame(null, $user->getBirthday());
         $this->assertNull($user->getCompanyId());
         $this->assertFalse($user->isVendor());
+        $this->assertSame('C', $user->getType());
 
         // shipping address
         $this->assertSame($userShipping->getTitle()->getValue(), $user->getShippingAddress()->getTitle()->getValue());
@@ -251,6 +253,7 @@ final class UserServiceTest extends ApiTestCase
         $this->assertSame(null, $user->getBirthday());
         $this->assertNull($user->getCompanyId());
         $this->assertFalse($user->isVendor());
+        $this->assertSame('C', $user->getType());
 
         //shipping adress
         $this->assertSame($userShipping->getTitle()->getValue(), $user->getShippingAddress()->getTitle()->getValue());
@@ -323,6 +326,7 @@ final class UserServiceTest extends ApiTestCase
         $this->assertSame('1998-07-12', $user->getBirthday()->format('Y-m-d'));
         $this->assertNull($user->getCompanyId());
         $this->assertFalse($user->isVendor());
+        $this->assertSame('C', $user->getType());
 
         // shipping address
         $this->assertTrue($addressCommand->getTitle()->equals($user->getShippingAddress()->getTitle()));
@@ -376,7 +380,7 @@ final class UserServiceTest extends ApiTestCase
         $this->assertSame(null, $user->getBirthday());
         $this->assertNull($user->getCompanyId());
         $this->assertFalse($user->isVendor());
-
+        $this->assertSame('C', $user->getType());
 
         $userService->updateUser(
             (new UpdateUserCommand())
@@ -398,6 +402,7 @@ final class UserServiceTest extends ApiTestCase
         $this->assertSame('1963-02-17', $user->getBirthday()->format('Y-m-d'));
         $this->assertNull($user->getCompanyId());
         $this->assertFalse($user->isVendor());
+        $this->assertSame('C', $user->getType());
     }
 
     public function testUpdateUserAddresses()
@@ -650,6 +655,7 @@ final class UserServiceTest extends ApiTestCase
         $this->assertSame('Jean', $user->getFirstname());
         $this->assertSame('Paul', $user->getLastname());
         $this->assertSame(null, $user->getBirthday());
+        $this->assertSame('C', $user->getType());
 
         $userService->updateUser(
             (new UpdateUserCommand())
@@ -667,6 +673,7 @@ final class UserServiceTest extends ApiTestCase
         $this->assertSame('Jacques', $user->getFirstname());
         $this->assertSame('Jules', $user->getLastname());
         $this->assertNull($user->getBirthday());
+        $this->assertSame('C', $user->getType());
     }
 
     public function testRecoverPassword()